Develop

+++ CodigoFonte/GPedU/BibliotecaViva/DTO/LocalizacaoGeograficaDTO.cs
@@ -2,7 +2,7 @@ namespace BibliotecaViva.DTO
 {
     public class LocalizacaoGeograficaDTO : BaseDTO
     {
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
+        public long Latitude { get; set; }
+        public long Longitude { get; set; }
     }
 }
